+++ src/Plivo/Resource/Application/ApplicationInterface.cs
@@ -43,6 +43,7 @@ namespace Plivo.Resource.Application
         /// <param name="defaultEndpointApp">Default endpoint app.</param>
         /// <param name="subaccount">Subaccount.</param>
         /// <param name="logIncomingMessages">Log incoming messages.</param>
+        /// <param name="publicUri">Public URI.</param>
 
         public ApplicationCreateResponse Create(
             string appName, string answerUrl = null, string answerMethod = null,
@@ -50,10 +51,11 @@ namespace Plivo.Resource.Application
             string fallbackAnswerUrl = null, string fallbackMethod = null,
             string messageUrl = null, string messageMethod = null,
             string defaultNumberApp = null, string defaultEndpointApp = null,
-            string subaccount = null, bool? logIncomingMessages = null)
+            string subaccount = null, bool? logIncomingMessages = null, bool? publicUri = null)
         {
             var mandatoryParams = new List<string> { "appName" }
                 ;
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -70,7 +72,9 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest,
+                    publicUri
                 });
 
 			return ExecuteWithExceptionUnwrap(() =>
@@ -98,6 +102,7 @@ namespace Plivo.Resource.Application
         /// <param name="defaultEndpointApp">Default endpoint app.</param>
         /// <param name="subaccount">Subaccount.</param>
         /// <param name="logIncomingMessages">Log incoming messages.</param>
+        /// <param name="publicUri">Public URI.</param>
 
         public async Task<ApplicationCreateResponse> CreateAsync(
             string appName, string answerUrl = null, string answerMethod = null,
@@ -105,10 +110,11 @@ namespace Plivo.Resource.Application
             string fallbackAnswerUrl = null, string fallbackMethod = null,
             string messageUrl = null, string messageMethod = null,
             string defaultNumberApp = null, string defaultEndpointApp = null,
-            string subaccount = null, bool? logIncomingMessages = null)
+            string subaccount = null, bool? logIncomingMessages = null, bool? publicUri = null)
         {
             var mandatoryParams = new List<string> {"appName"}
                 ;
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -125,7 +131,9 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest,
+                    publicUri
                 });
 
             var result = await Client.Update<ApplicationCreateResponse>(Uri, data);
@@ -144,7 +152,7 @@ namespace Plivo.Resource.Application
         {
 			return ExecuteWithExceptionUnwrap(() =>
 			{
-				var application = Task.Run(async () => await GetResource<Application>(appId).ConfigureAwait(false)).Result;
+				var application = Task.Run(async () => await GetResource<Application>(appId, new Dictionary<string, object> () { {"is_voice_request", true} }).ConfigureAwait(false)).Result;
 				application.Interface = this;
 				return application;
 			});
@@ -156,7 +164,7 @@ namespace Plivo.Resource.Application
         /// <param name="appId">App identifier.</param>
         public async Task<Application> GetAsync(string appId)
         {
-            var application = await GetResource<Application>(appId);
+            var application = await GetResource<Application>(appId, new Dictionary<string, object> () { {"is_voice_request", true} });
             application.Interface = this;
             return application;
         }
@@ -174,8 +182,9 @@ namespace Plivo.Resource.Application
             string subaccount = null, uint? limit = null, uint? offset = null)
         {
             var mandatoryParams = new List<string> { "" };
+            bool isVoiceRequest = true;
             var data = CreateData(
-                mandatoryParams, new { subaccount, limit, offset });
+                mandatoryParams, new { subaccount, limit, offset, isVoiceRequest });
 
 			return ExecuteWithExceptionUnwrap(() =>
 			{
@@ -198,8 +207,9 @@ namespace Plivo.Resource.Application
             string subaccount = null, uint? limit = null, uint? offset = null)
         {
             var mandatoryParams = new List<string> {""};
+            bool isVoiceRequest = true;
             var data = CreateData(
-                mandatoryParams, new {subaccount, limit, offset});
+                mandatoryParams, new {subaccount, limit, offset, isVoiceRequest});
 
             var resources = await ListResources<ListResponse<Application>>(data);
             resources.Objects.ForEach(
@@ -221,7 +231,8 @@ namespace Plivo.Resource.Application
         public DeleteResponse<Application> Delete(string appId, bool? cascade = null, string newEndpointApplication = null)
         {
             var data = new Dictionary<string, object> {};
-            data = CreateData(new List<string> {}, new {cascade, newEndpointApplication});
+            bool isVoiceRequest = true;
+            data = CreateData(new List<string> {}, new {cascade, newEndpointApplication, isVoiceRequest});
 
 			return ExecuteWithExceptionUnwrap(() =>
 			{
@@ -238,8 +249,8 @@ namespace Plivo.Resource.Application
         public async Task<DeleteResponse<Application>> DeleteAsync(string appId, bool? cascade = null, string newEndpointApplication = null)
         {
             var data = new Dictionary<string, object> { };
-            
-            data = CreateData(new List<string> { }, new { cascade, newEndpointApplication });
+            bool isVoiceRequest = true;
+            data = CreateData(new List<string> { }, new { cascade, newEndpointApplication, isVoiceRequest });
             return await DeleteResource<DeleteResponse<Application>>(appId, data);
         }
         #endregion
@@ -263,15 +274,18 @@ namespace Plivo.Resource.Application
         /// <param name="defaultEndpointApp">Default endpoint app.</param>
         /// <param name="subaccount">Subaccount.</param>
         /// <param name="logIncomingMessages">Log incoming messages.</param>
+        /// <param name="publicUri">Public URI.</param>
+
         public UpdateResponse<Application> Update(
             string appId, string answerUrl = null, string answerMethod = null,
             string hangupUrl = null, string hangupMethod = null,
             string fallbackAnswerUrl = null, string fallbackMethod = null,
             string messageUrl = null, string messageMethod = null,
             bool? defaultNumberApp = null, bool? defaultEndpointApp = null,
-            string subaccount = null, bool? logIncomingMessages = null)
+            string subaccount = null, bool? logIncomingMessages = null, bool? publicUri = null)
         {
             var mandatoryParams = new List<string> { "" };
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -287,7 +301,9 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest,
+                    publicUri
                 });
 
 			return ExecuteWithExceptionUnwrap(() =>
@@ -315,15 +331,18 @@ namespace Plivo.Resource.Application
         /// <param name="defaultEndpointApp">Default endpoint app.</param>
         /// <param name="subaccount">Subaccount.</param>
         /// <param name="logIncomingMessages">Log incoming messages.</param>
+        /// <param name="publicUri">Public URI.</param>
+
         public async Task<UpdateResponse<Application>> UpdateAsync(
             string appId, string answerUrl = null, string answerMethod = null,
             string hangupUrl = null, string hangupMethod = null,
             string fallbackAnswerUrl = null, string fallbackMethod = null,
             string messageUrl = null, string messageMethod = null,
             bool? defaultNumberApp = null, bool? defaultEndpointApp = null,
-            string subaccount = null, bool? logIncomingMessages = null)
+            string subaccount = null, bool? logIncomingMessages = null, bool? publicUri = null)
         {
             var mandatoryParams = new List<string> {""};
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -339,7 +358,9 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest,
+                    publicUri
                 });
 
             var result = await Client.Update<UpdateResponse<Application>>(Uri + appId + "/", data);