+++ src/Plivo/Resource/Powerpack/Powerpack.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Threading.Tasks;
 using Plivo.Resource.PhoneNumber;
-
+using System.Collections.Generic;
 namespace Plivo.Resource.Powerpack {
     /// <summary>
     /// Powerpack.
     /// </summary>
-
     public class Powerpack : Resource {
         public new string Id => uuid;
         public string number_pool_id { get; set; }
@@ -58,6 +57,12 @@ namespace Plivo.Resource.Powerpack {
         public string uuid { get; set; }
 
         /// <summary>
+        /// Number Priority
+        /// </summary>
+        /// <value>The number_priority .</value>
+        public List<NumberPriority> number_priority { get; set; }
+
+        /// <summary>
         /// Gets or sets the number_pool.
         /// </summary>
         /// <value>The number_pool.</value>
@@ -74,47 +79,47 @@ namespace Plivo.Resource.Powerpack {
         }
 
         public UpdateResponse<Powerpack> Update (string name = null, string application_type = null, string application_id = null,
-            bool? sticky_sender = null, bool? local_connect = null) {
+            bool? sticky_sender = null, bool? local_connect = null, List<NumberPriority> number_priority = null) {
             return ((PowerpackInterface) Interface)
-                .Update (Id, name, application_type, application_id, sticky_sender, local_connect);
+                .Update (Id, name, application_type, application_id, sticky_sender, local_connect, number_priority);
         }
         public async Task<UpdateResponse<Powerpack>> UpdateAsync (string name = null, string application_type = null, string application_id = null,
-            bool? sticky_sender = null, bool? local_connect = null) {
+            bool? sticky_sender = null, bool? local_connect = null, List<NumberPriority> number_priority = null) {
             return await ((PowerpackInterface) Interface)
-                .UpdateAsync (Id, name, application_type, application_id, sticky_sender, local_connect);
+                .UpdateAsync (Id, name, application_type, application_id, sticky_sender, local_connect, number_priority);
         }
         public ListResponse<Numbers> List_Numbers (string starts_with = null, string country_iso2 = null,
-            string type = null, uint? limit = null, uint? offset = null) {
+            string type = null, string service = null, uint? limit = null, uint? offset = null) {
             return ((PowerpackInterface) Interface)
-                .List_Numbers (number_pool_id, starts_with, country_iso2, type, limit, offset);
+                .List_Numbers (number_pool_id, starts_with, country_iso2, type, limit, offset, service);
         }
         public async Task<ListResponse<Numbers>> List_NumbersAsync (
             string starts_with = null, string country_iso2 = null,
-            string type = null, uint? limit = null, uint? offset = null) {
+            string type = null, string service = null, uint? limit = null, uint? offset = null) {
             return await ((PowerpackInterface) Interface)
-                .List_NumbersAsync (number_pool_id, starts_with, country_iso2, type, limit, offset);
+                .List_NumbersAsync (number_pool_id, starts_with, country_iso2, type, limit, offset, service);
 
         }
 
         public uint Count_Numbers (string starts_with = null, string country_iso2 = null,
-            string type = null, uint? limit = null, uint? offset = null) {
+            string type = null, string service = null, uint? limit = null, uint? offset = null) {
             return ((PowerpackInterface) Interface)
-                .Count_Number (number_pool_id, starts_with, country_iso2, type, limit, offset);
+                .Count_Number (number_pool_id, starts_with, country_iso2, type, limit, offset, service);
         }
         public async Task<uint> Count_NumbersAsync (
             string starts_with = null, string country_iso2 = null,
-            string type = null, uint? limit = null, uint? offset = null) {
+            string type = null, string service = null, uint? limit = null, uint? offset = null) {
             return await ((PowerpackInterface) Interface)
-                .Count_NumbersAsync (number_pool_id, starts_with, country_iso2, type, limit, offset);
+                .Count_NumbersAsync (number_pool_id, starts_with, country_iso2, type, limit, offset, service);
 
         }
-        public Numbers Add_Number (string number) {
+        public Numbers Add_Number (string number, string service = null) {
             return ((PowerpackInterface) Interface)
-                .Add_Number (number_pool_id, number);
+                .Add_Number (number_pool_id, number, service);
         }
-        public async Task<Numbers> Add_NumberAsync (string number) {
+        public async Task<Numbers> Add_NumberAsync (string number, string service = null) {
             return await ((PowerpackInterface) Interface)
-                .Add_NumberAsync (number_pool_id, number);
+                .Add_NumberAsync (number_pool_id, number, service);
         }
 
         public Tollfree Add_Tollfree (string tollfree) {
@@ -146,14 +151,14 @@ namespace Plivo.Resource.Powerpack {
                 .Remove_TollfreeAsync (number_pool_id, tollfree, unrent);
         }
 
-        public Numbers Find_Number (string number) {
+        public Numbers Find_Number (string number, string service = null) {
             return ((PowerpackInterface) Interface)
-                .Find_Number (number_pool_id, number);
+                .Find_Number (number_pool_id, number, service);
         }
         public async Task<Numbers> Find_NumberAsync (
-            string number) {
+            string number, string service = null) {
             return await ((PowerpackInterface) Interface)
-                .Find_NumberAsync (number_pool_id, number);
+                .Find_NumberAsync (number_pool_id, number, service);
         }
         public ListResponse<Shortcode> List_Shortcode (uint? limit = null, uint? offset = null) {
             return ((PowerpackInterface) Interface)
@@ -203,23 +208,23 @@ namespace Plivo.Resource.Powerpack {
                 .Find_TollfreeAsync (tollfree, number_pool_id);
         }
 
-        public Numbers Buy_Add_Number (string number = null, string type = null, string country_iso2 = null, string region = null, string pattern = null) {
+        public Numbers Buy_Add_Number (string number = null, string type = null, string country_iso2 = null, string region = null, string pattern = null, string service = null) {
             if (number == null) {
                 var countryIso = country_iso2;
                 var numbers = PhoneNumberI.List (countryIso, type, pattern, region);
                 number = numbers.Objects[0].Number;
             }
             return ((PowerpackInterface) Interface)
-                .Add_Number (number_pool_id, number, true);
+                .Add_Number (number_pool_id, number, service, true);
         }
-        public async Task<Numbers> Buy_Add_NumberAsync (string number = null, string type = null, string country_iso2 = null, string region = null, string pattern = null) {
+        public async Task<Numbers> Buy_Add_NumberAsync (string number = null, string type = null, string country_iso2 = null, string region = null, string pattern = null, string service = null) {
             if (number == null) {
                 var countryIso = country_iso2;
                 var numbers = PhoneNumberI.List (countryIso, type, pattern, region);
                 number = numbers.Objects[0].Number;
             }
             return await ((PowerpackInterface) Interface)
-                .Add_NumberAsync (number_pool_id, number, true);
+                .Add_NumberAsync (number_pool_id, number, service, true);
         }
 
         public override string ToString () {
@@ -231,7 +236,8 @@ namespace Plivo.Resource.Powerpack {
                 "ApplicationType: " + application_type + "\n" +
                 "ApplicationId: " + application_id + "\n" +
                 "NumberPool: " + number_pool + "\n" +
-                "CreatedOn: " + created_on + "\n";
+                "CreatedOn: " + created_on + "\n" +
+                "NumberPriority: " + number_priority + "\n";
         }
     }
 }