+++ src/Plivo/Resource/ResourceInterface.cs
@@ -108,7 +108,6 @@ namespace Plivo.Resource {
         public async Task<T> DeleteResource<T> (string id, Dictionary<string, object> data = null)
         where T : new () {
             var result = await Client.Delete<T> (Uri + id + "/", data);
-            Console.WriteLine (result.Object);
             try {
                 result.Object.GetType ().GetRuntimeProperty ("StatusCode").SetValue (result.Object, result.StatusCode, null);
             } catch (System.NullReferenceException) {
@@ -175,8 +174,12 @@ namespace Plivo.Resource {
                 return func ();
             } catch (AggregateException ex) {
                 ex.Flatten ();
+                if (ex.InnerExceptions[0] is Newtonsoft.Json.JsonReaderException){
+                    throw new PlivoValidationException ("Unexpected error occured. Please contact plivo support");
+                }
                 throw ex.InnerExceptions[0];
-            }
+                
+            } 
         }
 
         public static void ExecuteWithExceptionUnwrap (Action func) {